NAM2/azure_functions/requirements.txt
azure-functions
azure-cosmos
azure-storage-blob
azure-search-documents==11.6.0b9
azure-core
openai
requests
beautifulsoup4
python-dotenv
